requirements.txt
jsonlines==3.1.0
pandas==1.3.5
Pillow==9.4.0
seaborn==0.12.2
streamlit==1.20.0
